[LIVY-384] [Docs] Include local build of Docs
Include a copy of the most recent build of the Livy Docs in the website.
Once we make a Apache release we will replace these SNAPSHOT Docs with
the release Docs. Whether we include SNAPSHOT Docs in the future can be
discussed then. The process for updating the Docs can be found in the
Docs README in the main repo.