Improvements to getting started and community pages
I just heard about Livy the other day when it was announced it was included with [AWS EMR 5.9.0](https://aws.amazon.com/about-aws/whats-new/2017/10/support-for-apache-livy-hue-4-0-1-and-presto-0-184-on-amazon-emr-release-5-9-0/). I was checking out the website when evaluating how we could use Livy in our workflows and noticed a couple simple improvements I could make.

This PR changes the following on the Livy website:

* Adds a "Start Using" section to getting started page - provide a stronger call to action after you've set up Livy (including what the default port is)
* Improve grammar on the Community page - noticed a couple small tweaks to make here

Hopefully it's fine to provide both of these changes in the same PR.

Author: Devin Boyer <drb5272@gmail.com>

Closes #11 from drboyer/community+getting-started-improvements.